allow list now two words
