+++ dsf-bpe/dsf-bpe-process-update-allow-list/src/main/java/org/highmed/dsf/bpe/spring/config/UpdateAllowListConfig.java
@@ -1,0 +1,48 @@
+package org.highmed.dsf.bpe.spring.config;
+
+import org.camunda.bpm.engine.impl.cfg.ProcessEnginePlugin;
+import org.highmed.dsf.bpe.plugin.UpdateAllowListPlugin;
+import org.highmed.dsf.bpe.service.DownloadAllowList;
+import org.highmed.dsf.bpe.service.UpdateAllowList;
+import org.highmed.dsf.fhir.client.FhirWebserviceClientProvider;
+import org.highmed.dsf.fhir.organization.OrganizationProvider;
+import org.highmed.dsf.fhir.task.TaskHelper;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import ca.uhn.fhir.context.FhirContext;
+
+@Configuration
+public class UpdateAllowListConfig
+{
+	@Autowired
+	private FhirWebserviceClientProvider clientProvider;
+
+	@Autowired
+	private OrganizationProvider organizationProvider;
+
+	@Autowired
+	private TaskHelper taskHelper;
+
+	@Autowired
+	private FhirContext fhirContext;
+
+	@Bean
+	public ProcessEnginePlugin updateAllowListPlugin()
+	{
+		return new UpdateAllowListPlugin();
+	}
+
+	@Bean
+	public UpdateAllowList updateAllowList()
+	{
+		return new UpdateAllowList(organizationProvider, clientProvider, taskHelper);
+	}
+
+	@Bean
+	public DownloadAllowList downloadAllowList()
+	{
+		return new DownloadAllowList(clientProvider, taskHelper, fhirContext);
+	}
+}
