+++ dsf-bpe/dsf-bpe-process-update-allow-list/src/main/java/org/highmed/dsf/bpe/variables/ConstantsUpdateAllowList.java
@@ -1,0 +1,7 @@
+package org.highmed.dsf.bpe.variables;
+
+public interface ConstantsUpdateAllowList
+{
+	String CODESYSTEM_HIGHMED_UPDATE_ALLOW_LIST = "http://highmed.org/fhir/CodeSystem/update-allow-list";
+	String CODESYSTEM_HIGHMED_UPDATE_ALLOW_LIST_VALUE_ALLOW_LIST = "highmed_allow_list";
+}
