+++ dsf-bpe/dsf-bpe-process-update-allow-list/src/test/resources/log4j2.xml
