--- dsf-bpe/dsf-bpe-process-update-allowlist/pom.xml
@@ -1,24 +1,0 @@
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<artifactId>dsf-bpe-process-update-allowlist</artifactId>
-
-	<parent>
-		<groupId>org.highmed.dsf</groupId>
-		<artifactId>dsf-bpe-pom</artifactId>
-		<version>0.3.0-SNAPSHOT</version>
-	</parent>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.highmed.dsf</groupId>
-			<artifactId>dsf-bpe-process-base</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>de.hs-heilbronn.mi</groupId>
-			<artifactId>log4j2-utils</artifactId>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-</project>
