--- dsf-bpe/dsf-bpe-process-update-allowlist/src/main/java/org/highmed/dsf/bpe/variables/ConstantsUpdateAllowlist.java
@@ -1,7 +1,0 @@
-package org.highmed.dsf.bpe.variables;
-
-public interface ConstantsUpdateAllowlist
-{
-	String CODESYSTEM_HIGHMED_UPDATE_ALLOWLIST = "http://highmed.org/fhir/CodeSystem/update-allowlist";
-	String CODESYSTEM_HIGHMED_UPDATE_ALLOWLIST_VALUE_ALLOWLIST = "highmed_allowlist";
-}
