+++ dsf-fhir/dsf-fhir-server/src/main/resources/fhir/ActivityDefinition/downloadAllowList-0.3.0.xml.post
@@ -1,0 +1,1 @@
+url=http://highmed.org/bpe/Process/downloadAllowList&version=0.3.0
