--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/ActivityDefinition/downloadAllowlist-0.3.0.xml.post
@@ -1,1 +1,0 @@
-url=http://highmed.org/bpe/Process/downloadAllowlist&version=0.3.0
