--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/ActivityDefinition/updateAllowlist-0.3.0.xml
@@ -1,58 +1,0 @@
-<ActivityDefinition xmlns="http://hl7.org/fhir">
-	<meta>
-		<tag>
-			<system value="http://highmed.org/fhir/CodeSystem/authorization-role" />
-			<code value="REMOTE" />
-		</tag>
-	</meta>
-	<extension url="http://highmed.org/fhir/StructureDefinition/process-authorization">
-		<extension url="message-name">
-			<valueString value="updateAllowlistMessage" />
-		</extension>
-		<extension url="authorization-roles">
-			<extension url="authorization-role">
-				<valueCoding>
-					<system value="http://highmed.org/fhir/CodeSystem/authorization-role" />
-					<code value="LOCAL" />
-				</valueCoding>
-			</extension>
-		</extension>
-		<extension url="requester-organization-types">
-			<extension url="requester-organization-type">
-				<valueCoding>
-					<system value="http://highmed.org/fhir/CodeSystem/organization-type" />
-					<code value="TTP" />
-				</valueCoding>
-			</extension>
-		</extension>
-		<extension url="recipient-organization-types">
-			<extension url="recipient-organization-type">
-				<valueCoding>
-					<system value="http://highmed.org/fhir/CodeSystem/organization-type" />
-					<code value="TTP" />
-				</valueCoding>
-			</extension>
-		</extension>
-		<extension url="task-profile">
-			<valueCanonical value="http://highmed.org/fhir/StructureDefinition/highmed-task-update-allowlist" />
-		</extension>
-	</extension>
-	<url value="http://highmed.org/bpe/Process/updateAllowlist" />
-	<version value="0.3.0" />
-	<name value="UpdateAllowlist" />
-	<title value="Update Allowlist" />
-	<subtitle value="Update Allowlist Bundle Process" />
-	<status value="draft" />
-	<experimental value="true" />
-	<date value="2020-10-26" />
-	<publisher value="HiGHmed" />
-	<contact>
-		<name value="HiGHmed" />
-		<telecom>
-			<system value="email" />
-			<value value="pmo@highmed.org" />
-		</telecom>
-	</contact>
-	<description value="Process to update the allowlist FHIR bundle" />
-	<kind value="Task" />
-</ActivityDefinition>
