--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/ActivityDefinition/updateAllowlist-0.3.0.xml.post
@@ -1,1 +1,0 @@
-url=http://highmed.org/bpe/Process/updateAllowlist&version=0.3.0
