+++ dsf-fhir/dsf-fhir-server/src/main/resources/fhir/CodeSystem/update-allow-list-0.3.0.xml.post
@@ -1,0 +1,1 @@
+url=http://highmed.org/fhir/CodeSystem/update-allow-list&version=0.3.0
