--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/CodeSystem/update-allowlist-0.3.0.xml.post
@@ -1,1 +1,0 @@
-url=http://highmed.org/fhir/CodeSystem/update-allowlist&version=0.3.0
