--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/StructureDefinition/highmed-task-download-allowlist-0.3.0.xml
@@ -1,91 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<StructureDefinition xmlns="http://hl7.org/fhir">
-  <meta>
-    <tag>
-      <system value="http://highmed.org/fhir/CodeSystem/authorization-role" />
-      <code value="REMOTE" />
-    </tag>
-  </meta>
-  <url value="http://highmed.org/fhir/StructureDefinition/highmed-task-download-allowlist" />
-  <version value="0.3.0" />
-  <name value="TaskDownloadAllowlist" />
-  <status value="draft" />
-  <date value="2020-10-26" />
-  <fhirVersion value="4.0.1" />
-  <kind value="resource" />
-  <abstract value="false" />
-  <type value="Task" />
-  <baseDefinition value="http://highmed.org/fhir/StructureDefinition/highmed-task-base" />
-  <derivation value="constraint" />
-  <differential>
-    <element id="Task.instantiatesUri">
-      <path value="Task.instantiatesUri" />
-      <fixedUri value="http://highmed.org/bpe/Process/downloadAllowlist/0.3.0" />
-    </element>
-    <element id="Task.input">
-      <path value="Task.input" />
-      <min value="2" />
-      <max value="2" />
-    </element>
-    <element id="Task.input:message-name">
-      <path value="Task.input" />
-      <sliceName value="message-name" />
-    </element>
-    <element id="Task.input:message-name.value[x]">
-      <path value="Task.input.value[x]" />
-      <fixedString value="downloadAllowlistMessage" />
-    </element>
-    <element id="Task.input:business-key">
-      <path value="Task.input" />
-      <sliceName value="business-key" />
-      <max value="0" />
-    </element>
-    <element id="Task.input:correlation-key">
-      <path value="Task.input" />
-      <sliceName value="correlation-key" />
-      <max value="0" />
-    </element>
-    <element id="Task.input:bundle-reference">
-      <path value="Task.input" />
-      <sliceName value="bundle-reference" />
-      <min value="1" />
-      <max value="1" />
-    </element>
-    <element id="Task.input:bundle-reference.type">
-      <path value="Task.input.type" />
-      <binding>
-        <strength value="required" />
-        <valueSet value="http://highmed.org/fhir/ValueSet/update-allowlist" />
-      </binding>
-    </element>
-    <element id="Task.input:bundle-reference.type.coding">
-      <path value="Task.input.type.coding" />
-      <min value="1" />
-      <max value="1" />
-    </element>
-    <element id="Task.input:bundle-reference.type.coding.system">
-      <path value="Task.input.type.coding.system" />
-      <min value="1" />
-      <fixedUri value="http://highmed.org/fhir/CodeSystem/update-allowlist" />
-    </element>
-    <element id="Task.input:bundle-reference.type.coding.code">
-      <path value="Task.input.type.coding.code" />
-      <min value="1" />
-      <fixedCode value="highmed_allowlist" />
-    </element>
-    <element id="Task.input:bundle-reference.value[x]">
-      <path value="Task.input.value[x]" />
-      <type>
-        <code value="Reference" />
-      </type>
-    </element>
-    <element id="Task.input:bundle-reference.value[x].reference">
-      <path value="Task.input.value[x].reference" />
-      <min value="1" />
-    </element>
-    <element id="Task.input:bundle-reference.value[x].identifier">
-      <path value="Task.input.value[x].identifier" />
-      <max value="0" />
-    </element>
-  </differential>
-</StructureDefinition>
