--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/StructureDefinition/highmed-task-download-allowlist-0.3.0.xml.post
@@ -1,1 +1,0 @@
-url=http://highmed.org/fhir/StructureDefinition/highmed-task-download-allowlist&version=0.3.0
