--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/StructureDefinition/highmed-task-update-allowlist-0.3.0.xml
@@ -1,103 +1,0 @@
-<StructureDefinition xmlns="http://hl7.org/fhir">
-  <meta>
-    <tag>
-      <system value="http://highmed.org/fhir/CodeSystem/authorization-role" />
-      <code value="REMOTE" />
-    </tag>
-  </meta>
-  <url value="http://highmed.org/fhir/StructureDefinition/highmed-task-update-allowlist" />
-  <version value="0.3.0" />
-  <name value="TaskUpdateAllowlist" />
-  <status value="draft" />
-  <date value="2020-10-26" />
-  <fhirVersion value="4.0.1" />
-  <kind value="resource" />
-  <abstract value="false" />
-  <type value="Task" />
-  <baseDefinition value="http://highmed.org/fhir/StructureDefinition/highmed-task-base" />
-  <derivation value="constraint" />
-  <differential>
-    <element id="Task.instantiatesUri">
-      <path value="Task.instantiatesUri" />
-      <fixedUri value="http://highmed.org/bpe/Process/updateAllowlist/0.3.0" />
-    </element>
-    <element id="Task.input">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <max value="1" />
-    </element>
-    <element id="Task.input:message-name">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <sliceName value="message-name" />
-    </element>
-    <element id="Task.input:message-name.value[x]">
-      <path value="Task.input.value[x]" />
-      <fixedString value="updateAllowlistMessage" />
-    </element>
-    <element id="Task.input:business-key">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <sliceName value="business-key" />
-      <max value="0" />
-    </element>
-    <element id="Task.input:correlation-key">
-      <extension url="http://hl7.org/fhir/StructureDefinition/structuredefinition-explicit-type-name">
-        <valueString value="Parameter" />
-      </extension>
-      <path value="Task.input" />
-      <sliceName value="correlation-key" />
-      <max value="0" />
-    </element>
-    <element id="Task.output:allowlist">
-      <path value="Task.output" />
-      <sliceName value="allowlist" />
-      <max value="1" />
-    </element>
-    <element id="Task.output:allowlist.type">
-      <path value="Task.output.type" />
-      <binding>
-        <extension url="http://hl7.org/fhir/StructureDefinition/elementdefinition-bindingName">
-          <valueString value="TaskOutputParameterType" />
-        </extension>
-        <strength value="required" />
-        <valueSet value="http://highmed.org/fhir/ValueSet/update-allowlist" />
-      </binding>
-    </element>
-    <element id="Task.output:allowlist.type.coding">
-      <path value="Task.output.type.coding" />
-      <min value="1" />
-      <max value="1" />
-    </element>
-    <element id="Task.output:allowlist.type.coding.system">
-      <path value="Task.output.type.coding.system" />
-      <min value="1" />
-      <fixedUri value="http://highmed.org/fhir/CodeSystem/update-allowlist" />
-    </element>
-    <element id="Task.output:allowlist.type.coding.code">
-      <path value="Task.output.type.coding.code" />
-      <min value="1" />
-      <fixedCode value="highmed_allowlist" />
-    </element>
-    <element id="Task.output:allowlist.value[x]">
-      <path value="Task.output.value[x]" />
-      <type>
-        <code value="Reference" />
-      </type>
-    </element>
-    <element id="Task.output:allowlist.value[x].reference">
-      <path value="Task.output.value[x].reference" />
-      <min value="1" />
-    </element>
-    <element id="Task.output:allowlist.value[x].identifier">
-      <path value="Task.output.value[x].identifier" />
-      <max value="0" />
-    </element>
-  </differential>
-</StructureDefinition>
