+++ dsf-fhir/dsf-fhir-server/src/main/resources/fhir/ValueSet/update-allow-list-0.3.0.xml
@@ -1,0 +1,24 @@
+<ValueSet xmlns="http://hl7.org/fhir">
+	<meta>
+		<tag>
+			<system value="http://highmed.org/fhir/CodeSystem/authorization-role"/>
+			<code value="REMOTE"/>
+		</tag>		
+	</meta>
+	<url value="http://highmed.org/fhir/ValueSet/update-allow-list"/>
+	<version value="0.3.0"/>
+	<name value="HiGHmed_Update_Allow_List"/>
+	<title value="HiGHmed Update Allow List"/>
+	<status value="active"/>
+	<experimental value="false"/>
+	<date value="2020-10-26"/>
+	<publisher value="HiGHmed"/>
+	<description
+			value="ValueSet with standard values for the process update allow list"/>
+	<immutable value="true"/>
+	<compose>
+		<include>
+			<system value="http://highmed.org/fhir/CodeSystem/update-allow-list"/>
+		</include>
+	</compose>
+</ValueSet>
