--- dsf-fhir/dsf-fhir-server/src/main/resources/fhir/ValueSet/update-allowlist-0.3.0.xml.post
@@ -1,1 +1,0 @@
-url=http://highmed.org/fhir/ValueSet/update-allowlist&version=0.3.0
